requirements.txt
fastapi[standard]
pydantic[standard]
requests==2.28.1
uvicorn==0.24.0.post1
google==3.0.0
google-auth==2.40.3
google-genai==1.38.0
python-dotenv==1.2.1